type conversions.

+++ source/blender/makesrna/intern/rna_space.cc
@@ -976,12 +976,12 @@ static void rna_Space_show_region_ui_set(PointerRNA *ptr, bool value)
     ScrArea *area = rna_area_from_space(ptr);
     ARegion *region = BKE_area_find_region_type(area, RGN_TYPE_UI);
     if (BKE_regiontype_uses_category_tabs(region->runtime->type) &&
-        region->sizex <= UI_PANEL_CATEGORY_MIN_WIDTH)
+        float(region->sizex) <= UI_PANEL_CATEGORY_MIN_WIDTH)
     {
       /* If the region is showing only tabs, increase to full width. */
       const int new_width = region->runtime->type->prefsizex ? region->runtime->type->prefsizex :
-                                                               250.0f;
-      region->sizex = new_width;
+                                                               250;
+      region->sizex = short(new_width);
     }
   }
   rna_Space_bool_from_region_flag_set_by_type(ptr, RGN_TYPE_UI, RGN_FLAG_HIDDEN, !value);
